Move delete button to publishing accordion
