+++ src/pages/public_registration/ActionPanel.tsx
@@ -52,7 +52,8 @@ export const ActionPanel = ({
   const hasOtherPublishingRights =
     userHasAccessRight(registration, 'unpublish') ||
     userHasAccessRight(registration, 'republish') ||
-    userHasAccessRight(registration, 'terminate');
+    userHasAccessRight(registration, 'terminate') ||
+    userHasAccessRight(registration, 'delete');
 
   const customerHasConfiguredDoi = customer?.doiAgent.username;
   const canCreateDoiTicket =
@@ -70,10 +71,8 @@ export const ActionPanel = ({
     !!customerHasConfiguredDoi &&
     (canCreateDoiTicket || canApproveDoiTicket);
   const shouldSeeSupportAccordion = canCreateSupportTicket || canApproveSupportTicket;
-  const shouldSeeDelete = userHasAccessRight(registration, 'delete');
 
-  const canSeeTasksPanel =
-    shouldSeePublishingAccordion || shouldSeeDoiAccordion || shouldSeeSupportAccordion || shouldSeeDelete;
+  const canSeeTasksPanel = shouldSeePublishingAccordion || shouldSeeDoiAccordion || shouldSeeSupportAccordion;
 
   const [tabValue, setTabValue] = useState(canSeeTasksPanel ? TabValue.Tasks : TabValue.Log);
 
@@ -113,7 +112,6 @@ export const ActionPanel = ({
           shouldSeePublishingAccordion={shouldSeePublishingAccordion}
           shouldSeeDoiAccordion={shouldSeeDoiAccordion}
           shouldSeeSupportAccordion={shouldSeeSupportAccordion}
-          shouldSeeDelete={shouldSeeDelete}
           publishingRequestTickets={publishingRequestTickets}
           newestDoiRequestTicket={newestDoiRequestTicket}
           newestSupportTicket={newestSupportTicket}
