+++ src/pages/public_registration/action_accordions/MoreActionsCollapse.tsx
@@ -5,6 +5,8 @@ import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Registration } from '../../../types/registration.types';
 import { dataTestId } from '../../../utils/dataTestIds';
+import { userHasAccessRight } from '../../../utils/registration-helpers';
+import { DeleteDraftRegistration } from './DeleteDraftRegistration';
 import { RepublishRegistration } from './RepublishRegistration';
 import { TerminateRegistration } from './TerminateRegistration';
 import { UnpublishRegistration } from './UnpublishRegistration';
@@ -19,8 +21,9 @@ export const MoreActionsCollapse = ({ registration }: MoreActionsCollapseProps) 
 
   const isPublished = registration.status === 'PUBLISHED' || registration.status === 'PUBLISHED_METADATA';
   const isUnpublished = registration.status === 'UNPUBLISHED';
+  const canDeleteDraft = registration.status === 'DRAFT' && userHasAccessRight(registration, 'delete');
 
-  if (!(isPublished || isUnpublished)) {
+  if (!(isPublished || isUnpublished || canDeleteDraft)) {
     return null;
   }
 
@@ -45,6 +48,7 @@ export const MoreActionsCollapse = ({ registration }: MoreActionsCollapseProps) 
               <TerminateRegistration registration={registration} />
             </>
           )}
+          {canDeleteDraft && <DeleteDraftRegistration registration={registration} />}
         </Box>
       )}
     </Box>
